Add functionality on Home and Menu Pages
On frontend side, the menu page has been designed with advertising components and 3 products from the database.
Fixed bug where footer would position in the middle of the page if no product was fetched or if didn't have enough content on it.
Added loading animation on the menu, until products are fetched from the database.
Added the design of the login form.